Rename node-effect-cypher files to PascalCase
